# swagger.yaml
openapi: 3.0.0
info:
  title: TaskManagerSarmato
  description: RESTful API for task management
  version: 1.0.0
servers:
  - url: http://127.0.0.1:7777
paths:
  /api/login_check:
    post:
      summary: Authenticate User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: User's email address.
                  example: admin@mail.ru
                password:
                  type: string
                  description: User's password.
                  example: 12345
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Authentication token.
                    example: eyJ0e...Qwq7QDWy0ZzVQ
        '500':
          description: Server error
  /api/task/all:
    get:
      summary: Get list of tasks
      security:
        - JWT: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  responseCode:
                    type: integer
                    example: 200
                  data:
                    type: array
                    nullable: true
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Unique task ID
                          example: 1
                        title:
                          type: string
                          description: Title task
                          example: Title task
                        description:
                          type: string
                          description: Description task
                          example: Description task
                        createdAt:
                          type: integer
                          description: Date the task was created (in UNIX timestamp format)
                          example: 1700170901
                        updatedAt:
                          type: integer
                          nullable: true
                          description: >-
                            The date of the last update of the task (in UNIX
                            timestamp format)
                          example: 1700170901
                        planeCompleteDate:
                          type: integer
                          description: >-
                            Planned task completion date (in UNIX timestamp
                            format)
                          example: 1700170901
                        isComplete:
                          type: boolean
                          description: Task status (false - in progress, true - completed)
                          example: 0
                        userId:
                          type: integer
                          description: ID of the user associated with the task
                          example: 1
        '500':
          description: Server error
  /api/task:
    get:
      summary: Get task by ID
      security:
        - JWT: []
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
          description: Unique task ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  responseCode:
                    type: integer
                    example: 200
                  data:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: integer
                        description: Unique task ID
                        example: 1
                      title:
                        type: string
                        description: Title task
                        example: Title task
                      description:
                        type: string
                        description: Description task
                        example: Description task
                      createdAt:
                        type: integer
                        description: Date the task was created (in UNIX timestamp format)
                        example: 1700170901
                      updatedAt:
                        type: integer
                        nullable: true
                        description: >-
                          The date of the last update of the task (in UNIX
                          timestamp format)
                        example: 1700170901
                      planeCompleteDate:
                        type: integer
                        description: >-
                          Planned task completion date (in UNIX timestamp
                          format)
                        example: 1700170901
                      isComplete:
                        type: boolean
                        description: Task status (false - in progress, true - completed)
                        example: 0
                      userId:
                        type: integer
                        description: ID of the user associated with the task
                        example: 1
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
        '401':
          description: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
    post:
      summary: Create a new task
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title task
                  example: Title task
                description:
                  type: string
                  description: Description task
                  example: Description task
                planeCompleteDate:
                  type: integer
                  description: Planned task completion date (in UNIX timestamp format)
                  example: 1700170901
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  responseCode:
                    type: integer
                    example: 200
                  data:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: integer
                        description: Unique task ID
                        example: 1
                      title:
                        type: string
                        description: Title task
                        example: Title task
                      description:
                        type: string
                        description: Description task
                        example: Description task
                      createdAt:
                        type: integer
                        description: Date the task was created (in UNIX timestamp format)
                        example: 1700170901
                      updatedAt:
                        type: integer
                        nullable: true
                        description: >-
                          The date of the last update of the task (in UNIX
                          timestamp format)
                        example: 1700170901
                      planeCompleteDate:
                        type: integer
                        description: >-
                          Planned task completion date (in UNIX timestamp
                          format)
                        example: 1700170901
                      isComplete:
                        type: boolean
                        description: Task status (false - in progress, true - completed)
                        example: 0
                      userId:
                        type: integer
                        description: ID of the user associated with the task
                        example: 1
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
        '401':
          description: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
    patch:
      summary: Update a task
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: Unique task ID
                  example: 1
                title:
                  type: string
                  description: Title task
                  example: Title task
                description:
                  type: string
                  description: Description task
                  example: Description task
                planeCompleteDate:
                  type: integer
                  description: Planned task completion date (in UNIX timestamp format)
                  example: 1700170901
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  responseCode:
                    type: integer
                    example: 200
                  data:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: integer
                        description: Unique task ID
                        example: 1
                      title:
                        type: string
                        description: Title task
                        example: Title task
                      description:
                        type: string
                        description: Description task
                        example: Description task
                      createdAt:
                        type: integer
                        description: Date the task was created (in UNIX timestamp format)
                        example: 1700170901
                      updatedAt:
                        type: integer
                        nullable: true
                        description: >-
                          The date of the last update of the task (in UNIX
                          timestamp format)
                        example: 1700170901
                      planeCompleteDate:
                        type: integer
                        description: >-
                          Planned task completion date (in UNIX timestamp
                          format)
                        example: 1700170901
                      isComplete:
                        type: boolean
                        description: Task status (false - in progress, true - completed)
                        example: 0
                      userId:
                        type: integer
                        description: ID of the user associated with the task
                        example: 1
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
        '401':
          description: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
    delete:
      summary: Delete a task
      security:
        - JWT: []
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
          description: Unique task ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
        '401':
          description: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
  /api/task/done:
    patch:
      summary: Mark the task as completed
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: Unique task ID
                  example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  responseCode:
                    type: integer
                    example: 200
                  data:
                    type: object
                    nullable: true
                    properties:
                      id:
                        type: integer
                        description: Unique task ID
                        example: 1
                      title:
                        type: string
                        description: Title task
                        example: Title task
                      description:
                        type: string
                        description: Description task
                        example: Description task
                      createdAt:
                        type: integer
                        description: Date the task was created (in UNIX timestamp format)
                        example: 1700170901
                      updatedAt:
                        type: integer
                        nullable: true
                        description: >-
                          The date of the last update of the task (in UNIX
                          timestamp format)
                        example: 1700170901
                      planeCompleteDate:
                        type: integer
                        description: >-
                          Planned task completion date (in UNIX timestamp
                          format)
                        example: 1700170901
                      isComplete:
                        type: boolean
                        description: Task status (false - in progress, true - completed)
                        example: 0
                      userId:
                        type: integer
                        description: ID of the user associated with the task
                        example: 1
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
        '401':
          description: Unauthorized
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    nullable: true
                    example: text error
                  data:
                    type: object
                    nullable: true
                  responseCode:
                    type: integer
                    example: 200
components:
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization